add linux stub so tests run
